LPS-57119 Move classes that are still sitting in portal\tools\ servicebuilder and portal\tools\sourceformatter that are being used elsewhere to portal\tools
